Riorganizzazione codice: spostate funzioni non cardini da utente.c a utile_utente.c
Spostate funzioni non cardini dell’ADT Utente in un modulo separato per migliorare la separazione delle responsabilità
Nessuna modifica al codice delle funzioni, solo riorganizzazione file
